Rename libwebviewchromium.so to libwebviewchromium2.so
This allows the native library to be loaded on Kitkat,
where there's already a system library with the same name.

This makes the package incompatible with standard WebView loader
on Lollipop (it doesn't read the library name from manifest).
You'd have to inject the provider manually instead of overriding
`android:string/config_webViewPackageName` at run time.